requirements.txt
matplotlib==3.5.2
py_cpuinfo==8.0.0
scipy==1.8.1
tqdm==4.64.0
torch==1.11.0
